+++ transformer_engine/pytorch/sequential/nvte/activation.py
@@ -1,5 +1,5 @@
 from __future__ import annotations
-from .. import cpp_extensions as _nvte
+from . import cpp_extensions as _nvte
 from .empty import empty
 from ._common import torch_op
