+++ transformer_engine/pytorch/sequential/nvte/dtype.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
-from .. import cpp_extensions as _nvte
-from ..cpp_extensions import te_to_torch_dtype, torch_to_te_dtype, dtype_name, bit_width
+from . import cpp_extensions as _nvte
+from .cpp_extensions import te_to_torch_dtype, torch_to_te_dtype, dtype_name, bit_width
 
 
 def is_fp8(t: _nvte.Tensor | _nvte.DType):
