+++ transformer_engine/pytorch/sequential/nvte/misc_fusions.py
@@ -1,5 +1,5 @@
 from __future__ import annotations
-from .. import cpp_extensions as _nvte
+from . import cpp_extensions as _nvte
 from ._common import torch_op
 from .dtype import is_fp8
 from .cast_transpose import cast_transpose_checked
